refactor(fossid-webapp): Extract the function `mapSnippetFindings`
Extracting the function improves the readability of the `FossId` class
and the `createResultSummary` function. It also allows adding tests for
the `mapSnippetFindings` function in the future.
