+++ scanner/src/main/kotlin/scanners/fossid/FossIdScanResults.kt
@@ -24,12 +24,22 @@ import org.ossreviewtoolkit.clients.fossid.model.identification.ignored.IgnoredF
 import org.ossreviewtoolkit.clients.fossid.model.identification.markedAsIdentified.MarkedAsIdentifiedFile
 import org.ossreviewtoolkit.clients.fossid.model.result.Snippet
 import org.ossreviewtoolkit.clients.fossid.model.summary.Summarizable
+import org.ossreviewtoolkit.model.ArtifactProvenance
 import org.ossreviewtoolkit.model.CopyrightFinding
+import org.ossreviewtoolkit.model.Hash
 import org.ossreviewtoolkit.model.Issue
 import org.ossreviewtoolkit.model.LicenseFinding
+import org.ossreviewtoolkit.model.PackageProvider
+import org.ossreviewtoolkit.model.RemoteArtifact
+import org.ossreviewtoolkit.model.Snippet as OrtSnippet
+import org.ossreviewtoolkit.model.SnippetFinding
 import org.ossreviewtoolkit.model.TextLocation
+import org.ossreviewtoolkit.model.config.ScannerConfiguration
 import org.ossreviewtoolkit.model.createAndLogIssue
+import org.ossreviewtoolkit.model.utils.PurlType
 import org.ossreviewtoolkit.utils.common.collectMessages
+import org.ossreviewtoolkit.utils.spdx.SpdxConstants
+import org.ossreviewtoolkit.utils.spdx.toSpdx
 
 /**
  * A data class to hold FossID raw results.
@@ -92,3 +102,97 @@ internal fun <T : Summarizable> List<T>.mapSummary(
         copyrightFindings = copyrightFindings
     )
 }
+
+/**
+ * Map the raw snippets to ORT [SnippetFinding]s to be included in the [ScanSummary]. If a snippet license cannot be
+ * parsed, an issues is added to [issues].
+ */
+internal fun mapSnippetFindings(
+    rawResults: RawResults,
+    scannerConfig: ScannerConfiguration,
+    issues: MutableList<Issue>
+): Set<SnippetFinding> {
+    val fakeLocation = TextLocation(".", TextLocation.UNKNOWN_LINE)
+
+    return rawResults.listSnippets.flatMap { (file, rawSnippets) ->
+        val snippets = rawSnippets.map {
+            val license = it.artifactLicense?.let {
+                runCatching {
+                    LicenseFinding.createAndMap(
+                        it,
+                        fakeLocation,
+                        detectedLicenseMapping = scannerConfig.detectedLicenseMapping
+                    ).license
+                }.onFailure { spdxException ->
+                    issues += FossId.createAndLogIssue(
+                        source = "FossId",
+                        message = "Failed to parse license '$it' as an SPDX expression:" +
+                                " ${spdxException.collectMessages()}"
+                    )
+                }.getOrNull()
+            } ?: SpdxConstants.NOASSERTION.toSpdx()
+
+            // FossID does not return the hash of the remote artifact. Instead, it returns the MD5 hash of the
+            // matched file in the remote artifact as part of the "match_file_id" property.
+            val url = checkNotNull(it.url) {
+                "The URL of snippet ${it.id} must not be null."
+            }
+            val snippetProvenance = ArtifactProvenance(RemoteArtifact(url, Hash.NONE))
+            val purl = it.purl ?: "pkg:${urlToPackageType(url)}/${it.author}/${it.artifact}@${it.version}"
+
+            val additionalSnippetData = mapOf(
+                FossId.SNIPPET_DATA_ID to it.id.toString(),
+                FossId.SNIPPET_DATA_MATCH_TYPE to it.matchType.toString(),
+                FossId.SNIPPET_DATA_RELEASE_DATE to it.releaseDate.orEmpty()
+            )
+
+            // TODO: FossID doesn't return the line numbers of the match, only the character range. One must use
+            //       another call "getMatchedLine" to retrieve the matched line numbers. Unfortunately, this is a
+            //       call per snippet which is too expensive. When it is available for a batch of snippets, it can
+            //       be used here.
+            OrtSnippet(
+                it.score.toFloat(),
+                TextLocation(it.file, TextLocation.UNKNOWN_LINE),
+                snippetProvenance,
+                purl,
+                license,
+                additionalSnippetData
+            )
+        }
+
+        val sourceLocation = TextLocation(file, TextLocation.UNKNOWN_LINE)
+        snippets.map {
+            SnippetFinding(
+                sourceLocation,
+                it
+            )
+        }
+    }.toSet()
+}
+
+/**
+ * Return the [PurlType] as determined from the given [url], or [PurlType.GENERIC] if there is no match.
+ */
+private fun urlToPackageType(url: String): PurlType =
+    when (val provider = PackageProvider.get(url)) {
+        PackageProvider.COCOAPODS -> PurlType.COCOAPODS
+        PackageProvider.CRATES_IO -> PurlType.CARGO
+        PackageProvider.DEBIAN -> PurlType.DEBIAN
+        PackageProvider.GITHUB -> PurlType.GITHUB
+        PackageProvider.GITLAB -> PurlType.GITLAB
+        PackageProvider.GOLANG -> PurlType.GOLANG
+        PackageProvider.MAVEN_CENTRAL, PackageProvider.MAVEN_GOOGLE -> PurlType.MAVEN
+        PackageProvider.NPM_JS -> PurlType.NPM
+        PackageProvider.NUGET -> PurlType.NUGET
+        PackageProvider.PACKAGIST -> PurlType.COMPOSER
+        PackageProvider.PYPI -> PurlType.PYPI
+        PackageProvider.RUBYGEMS -> PurlType.GEM
+
+        else -> {
+            PurlType.GENERIC.also {
+                FossId.logger.warn {
+                    "Cannot determine PURL type for url '$url' and provider '$provider'. Falling back to '$it'."
+                }
+            }
+        }
+    }
